Peformance boost from removing header

+++ include/mv/macro.h
@@ -8,7 +8,7 @@
 
 #define MV_INLINE static inline __attribute__((always_inline))
 #define mv_make_key(r, t) ((((uint64_t)(r) << 32) | (uint64_t)(t)))
-#define mv_make_rdz_key(x, y) mv_make_key(x, ((1 << 31) | y));
+#define mv_make_rdz_key(x, y) mv_make_key(x, (((uint32_t) 1 << 31) | y));
 #define likely(x) __builtin_expect(!!(x), 1)
 #define unlikely(x) __builtin_expect(!!(x), 0)
